custom_skill/requirements.txt
azure-functions
azure-identity
azure-storage-blob
scikit-learn
pandas
numpy
requests
python-dotenv
